+++ src/views/nodes/compareResultsNode.ts
@@ -318,9 +318,15 @@ export class CompareResultsNode extends ViewNode<SearchAndCompareView> {
 		}
 
 		const files = (await this.view.container.git.getDiffStatus(this.repoPath, comparison)) ?? [];
+		const shortstat = await this.view.container.git.getChangedFilesCount(this.uri.repoPath!, comparison);
 
 		return {
-			label: `${pluralize('file', files.length, { zero: 'No' })} changed`,
+			label:
+				`${pluralize('file', files.length, { zero: 'No' })} changed` +
+				` with ${pluralize('addition', shortstat?.additions ?? 0)} and ${pluralize(
+					'deletion',
+					shortstat?.deletions ?? 0,
+				)}`,
 			files: files,
 		};
 	}